Вынесла функции в отдельные модули
