+++ public/index.html
@@ -26,6 +26,6 @@
   </section>
 </footer>
 
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
