+++ src/utils.js
@@ -1,0 +1,3 @@
+const render = (container, template, place) => container.insertAdjacentHTML(place, template);
+
+export {render};
