+++ src/view/load-more-btn.js
@@ -1,0 +1,1 @@
+export const createLoadMoreBtnTemplate = () => `<button class="films-list__show-more">Show more</button>`;
